step_add function maximum up

+++ step.c
@@ -2,10 +2,10 @@
 //It's have a CPlusPlus edition
 
 
-long step_add(long max)  //max must be float
+long long step_add(long long max)  //max must be float
 {	
- 	long j=0.0; //define j
- 	for (long i=0.0;i<max+1;i+=1)  //a easy for loop
+ 	long long j=0; //define j
+ 	for (long long i=0;i<max+1;i+=1)  //a easy for loop
  	{
  		j+=i;  //Plus
  	}
